DAT-807: rename llm/tool_repair.py -> llm/contract_repair.py
The module now holds exactly one function, repair_tool_contract, which parses
the emitted SQL and checks the declared provenance against the served relation
schema. It never repaired a tool payload; its schema twin (repair_tool_output)
died with the forced tool.
